logweather/requirements.txt
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
idna==2.10
mysql-connector==2.2.9
numpy==1.20.1
pandas==1.2.2
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
zope.interface==5.2.0
